Compactando codigo Modals
