+++ src/views/pages/SupportClients.js
@@ -35,6 +35,8 @@ import {
 
 // core components
 import ReactTable from "components/ReactTable/ReactTable.js";
+import ModalUpdateSupport from "../components/modals/ModalUpdateSupport.js";
+import ModalReadSupport from "../components/modals/ModalReadSupport.js";
 
 const dataTable = [
   ["Tiger Nixon", "System Architect", "Edinburgh", "61"],
@@ -161,14 +163,6 @@ function SupportClients(props) {
     //Aqui vamos a descargar la lista de clientes de la base de datos por primera vez
   }, []);
 
-  function addRecord(event) {
-    //Código para añadir un registro a la tabla
-    //EndPoint CREATE
-
-    //una vez que añadimos el nuevo usuario, vamos a actualizar la tabla
-    //updateRecords();
-  }
-
   function updateRecord(){
     //A la hora de crear un nuevo registro necesitamos actualizar la tabla para que
     //se pinten todos los registros incluido el nuevo
@@ -179,15 +173,6 @@ function SupportClients(props) {
   function readRecord(){
     //Leemos la informacion completa del registo para pintarla en el modal
     //tal vez no sea necesaria porque ya se leyó anteriormente...
-  }
-
-  function toggleModalAddRecord(){
-    if(modalAddRecord == false){
-      setModalAddRecord(true);
-    }
-    else{
-      setModalAddRecord(false);
-    }
   }
 
   function toggleModalReadRecord(){
@@ -267,68 +252,11 @@ function SupportClients(props) {
         </Row>
       </div>
 
-      {/*MODAL PARA AÑADIR REGISTROS*/}
-      <Modal isOpen={modalAddRecord} toggle={toggleModalAddRecord}>
-        <div className="modal-header justify-content-center">
-        <button type="button" className="close" data-dismiss="modal" aria-label="Close" onClick={toggleModalAddRecord}>
-            <span aria-hidden="true">×</span>
-        </button>
-        <h5 className="modal-title">Add new record</h5>
-        </div>
-        <ModalBody>
-            <p>Woohoo, you're reading this text in a modal!</p>
-        </ModalBody>
-        <ModalFooter>
-            <Button color="secondary" onClick={toggleModalAddRecord}>
-                Close
-            </Button>
-            <Button color="primary">
-                Save changes
-            </Button>
-        </ModalFooter>
-      </Modal>
-
       {/*MODAL PARA LEER REGISTRO*/}
-      <Modal isOpen={modalReadRecord} toggle={toggleModalReadRecord}>
-        <div className="modal-header justify-content-center">
-        <button type="button" className="close" data-dismiss="modal" aria-label="Close" onClick={toggleModalReadRecord}>
-            <span aria-hidden="true">×</span>
-        </button>
-        <h5 className="modal-title">Record Detail</h5>
-        </div>
-        <ModalBody>
-            <p>Woohoo, you're reading this text in a modal!</p>
-        </ModalBody>
-        <ModalFooter>
-            <Button color="secondary" onClick={toggleModalReadRecord}>
-                Close
-            </Button>
-            <Button color="primary">
-                Save changes
-            </Button>
-        </ModalFooter>
-      </Modal>
+      <ModalReadSupport abierto = {modalReadRecord} toggleModalReadRecord = {toggleModalReadRecord}/>
 
       {/*MODAL PARA MODIFICAR REGISTRO*/}
-      <Modal isOpen={modalUpdateRecord} toggle={toggleModalUpdateRecord}>
-        <div className="modal-header justify-content-center">
-        <button type="button" className="close" data-dismiss="modal" aria-label="Close" onClick={toggleModalUpdateRecord}>
-            <span aria-hidden="true">×</span>
-        </button>
-        <h5 className="modal-title">Edit Record</h5>
-        </div>
-        <ModalBody>
-            <p>Woohoo, you're reading this text in a modal!</p>
-        </ModalBody>
-        <ModalFooter>
-            <Button color="secondary" onClick={toggleModalUpdateRecord}>
-                Close
-            </Button>
-            <Button color="primary">
-                Save changes
-            </Button>
-        </ModalFooter>
-      </Modal>
+      <ModalUpdateSupport abierto = {modalUpdateRecord} toggleModalUpdateRecord = {toggleModalUpdateRecord}/>
     </>
   );
 }
